Name change: application --> renterProfile
