--- controllers/users/renter/applicationController.js
@@ -1,29 +1,0 @@
-let Services = require('../../../services/users/renter/applicationService');
-
-const createApplication = (req, res) => { 
-    let appData = req.query;
-    if (Services.createApplication(appData)){
-        res.status(200).send();
-    }
-    else {
-        res.status(400).send();
-    } 
-   
-}
-
-
-const deleteApplication = (req, res) => { 
-    let userId = req.query.userId;
-    if (Services.deleteApplication(userId)){
-        res.status(200).send();
-    }
-    else {
-        res.status(400).send();
-    } 
-   
-}
-
-
-
-
-module.exports = {createApplication, deleteApplication}
