--- routes/users/renter/application.js
@@ -1,17 +1,0 @@
-var express = require('express');
-var router = express.Router();
-var Controller = require ('../../../controllers/users/renter/applicationController.js');
-
-
-
-router.post('/', (req, res) => {
-    Controller.createApplication(req, res);
-});
-
-router.delete('/', (req, res) => {
-    Controller.deleteApplication(req, res)
-})
-
-
-module.exports = router;
-
