+++ routes/users/renter/renterProfile.js
@@ -1,0 +1,17 @@
+var express = require('express');
+var router = express.Router();
+var Controller = require ('../../../controllers/users/renter/renterProfileController.js');
+
+
+
+router.post('/', (req, res) => {
+    Controller.createApplication(req, res);
+});
+
+router.delete('/', (req, res) => {
+    Controller.deleteApplication(req, res)
+})
+
+
+module.exports = router;
+
