+++ server.js
@@ -22,11 +22,11 @@ app.use(express.static(__dirname + '/public'));
 
 let userRoute = require('./routes/users/user');
 let propertyRoute = require('./routes/users/landlord/property');
-let rentalApplicationRoute = require('./routes/users/renter/application');
+let rentalApplicationRoute = require('./routes/users/renter/renterProfile');
 
 app.use('/user', userRoute);
 app.use('/property', propertyRoute);
-app.use('/application', rentalApplicationRoute);
+app.use('/renterProfile', rentalApplicationRoute);
